[SWM-290] Refactor : ui component directory
